Fix Team City regression test breakage.
## Rationale:

Adding **uiautomator** resulted in breakage on Team City regression testing.  The issue is that the BasicOnceUxSampleApp now runs UIAutomator tests rather than integration tests (which have been moved to the ANP build).  But **uiatomator** now is driven by *Ant* rather than *Gradle* and attempting to use *Gradle* to run the **connectedCheck** task will result in compilation errors and a runtime exception.  The fix is to disable the task by redefining it to be a no-op.

## Changed files:

BasicOnceUxSampleApp/build.gradle
* Redefine the **connectedCheck** task.